Communication errors triggers slot update using the async api
When a command response indicates a communication error the slot map
is updated. This is now updated using the async api, which avoids
blocking calls to connect when querying the cluster node.

A random already connected node is primarily used, but any random node
is used if none are already connected.
cc->update_route_time is used to throttle the slotmap update and to make
sure we only have one ongoing query at a time (SLOTMAP_UPDATE_ONGOING).